requirement.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-crontab==0.7.1
django-filter==21.1
djangorestframework==3.11.1
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
